requirements.txt
discord.py-self==1.9.2
aiohttp==3.7.4.post0
requests
legacy-cgi
pynacl
